refactor(domain): split lecture duration resolver support
